requirements-dev.txt
pytest>=7.0.0
pyyaml>=6.0
